refactor: extract SettingsContent form logic into hooks
